Add a new noclip object type and the related controller script
