+++ Assets/Code/Scripts/NoclipObjectControllers.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 608ce688333a47798333bdb4604dfbc0
+timeCreated: 1667511532
