+++ Assets/Code/Scripts/NoclipObjectControllers/BaseNoclipObjectController.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 7e4c652ccd9f41f790f62e5594b0ef73
+timeCreated: 1667511554
