+++ Assets/Code/Scripts/NoclipObjectControllers/NoclipIntangibleController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 306aee0b1f14d874191d250efd5dd1bc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
