+++ Assets/Level/Prefabs/NoclipIntangiblePlatform.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d8c021965fbd8ba468562529de84ed0d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
